requirements.txt
django-oscar==0.5.1
requests==2.0.1

